src/models/mobilenet/mobilenet.py
from typing import Callable, List, Optional, Sequence, Tuple, Type, Union

import torch
import torch.nn as nn
from torch import Tensor
from torchvision.models._utils import _make_divisible
from torchvision.utils import _make_ntuple

import src.sparse as sparse
from src.models.modeling_utils import apply_bn_to_conv_out, choose_correct_bn_type, create_general_conv2d

BNLayer = Union[nn.BatchNorm2d, sparse.MaskedBatchNorm2d]
Conv2dLayer = Union[
    nn.Conv2d,
    sparse.MaskedConv2d,
]


class Conv2dNormActivation(nn.Module):
    """We change the original ConvNormActivation to support MaskedBatchNorm2d by making it a nn.Module
    instead of a nn.Sequential. This was important because the BN layer had different signatures for Masked
     vs not masked versions - cf apply_bn_to_conv_out() in utils.py.
    """

    def __init__(
        self,
        in_channels: int,
        out_channels: int,
        kernel_size: Union[int, Tuple[int, ...]] = 3,
        stride: Union[int, Tuple[int, ...]] = 1,
        padding: Optional[Union[int, Tuple[int, ...], str]] = None,
        groups: int = 1,
        norm_layer: Type[BNLayer] = nn.BatchNorm2d,
        activation_layer: Optional[Callable[..., torch.nn.Module]] = nn.ReLU,
        dilation: Union[int, Tuple[int, ...]] = 1,
        inplace: Optional[bool] = True,
        bias: Optional[bool] = None,
        conv_layer: Type[Conv2dLayer] = nn.Conv2d,
    ) -> None:

        super().__init__()

        if padding is None:
            if isinstance(kernel_size, int) and isinstance(dilation, int):
                padding = (kernel_size - 1) // 2 * dilation
            else:
                _conv_dim = len(kernel_size) if isinstance(kernel_size, Sequence) else len(dilation)
                kernel_size = _make_ntuple(kernel_size, _conv_dim)
                dilation = _make_ntuple(dilation, _conv_dim)
                padding = tuple((kernel_size[i] - 1) // 2 * dilation[i] for i in range(_conv_dim))
        if bias is None:
            bias = norm_layer is None

        self.conv1 = create_general_conv2d(
            conv_layer=conv_layer,
            in_channels=in_channels,
            out_channels=out_channels,
            kernel_size=kernel_size,
            stride=stride,
            padding=padding,
            dilation=dilation,
            groups=groups,
            use_bias=bias,
            masked_layer_kwargs={},
        )

        bn1_type = choose_correct_bn_type(self.conv1, norm_layer)
        self.bn1 = bn1_type(num_features=out_channels)

        params = {} if inplace is None else {"inplace": inplace}
        self.activation = activation_layer(**params)

        self.out_channels = out_channels

    def forward(self, x: Tensor) -> Tensor:
        out = self.conv1(x)
        out = apply_bn_to_conv_out(self.bn1, out)
        out = self.activation(out)

        return out


# necessary for backwards compatibility
class InvertedResidual(nn.Module):
    def __init__(
        self,
        inp: int,
        oup: int,
        stride: int,
        expand_ratio: int,
        norm_layer: Type[BNLayer] = nn.BatchNorm2d,
        conv_layer: Type[Conv2dLayer] = nn.Conv2d,
    ) -> None:
        super().__init__()
        self.stride = stride
        if stride not in [1, 2]:
            raise ValueError(f"stride should be 1 or 2 instead of {stride}")

        hidden_dim = int(round(inp * expand_ratio))
        self.use_res_connect = self.stride == 1 and inp == oup

        layers: List[nn.Module] = []
        if expand_ratio != 1:
            # pw
            layers.append(
                Conv2dNormActivation(
                    inp,
                    hidden_dim,
                    kernel_size=1,
                    norm_layer=norm_layer,
                    activation_layer=nn.ReLU6,
                    conv_layer=conv_layer,
                )
            )
        layers.extend(
            [
                # depthwise
                Conv2dNormActivation(
                    hidden_dim,
                    hidden_dim,
                    stride=stride,
                    groups=hidden_dim,
                    norm_layer=norm_layer,
                    activation_layer=nn.ReLU6,
                    conv_layer=conv_layer,
                ),
                # piecewise-linear
                create_general_conv2d(
                    conv_layer=conv_layer,
                    in_channels=hidden_dim,
                    out_channels=oup,
                    kernel_size=1,
                    stride=1,
                    padding=0,
                    dilation=1,
                    groups=1,
                    use_bias=False,
                    masked_layer_kwargs={},
                ),
            ]
        )
        self.conv = nn.Sequential(*layers)
        self.bn = norm_layer(num_features=oup)
        self.out_channels = oup
        self._is_cn = stride > 1

    def forward(self, x: Tensor) -> Tensor:
        out = self.conv(x)
        out = apply_bn_to_conv_out(self.bn, out)
        if self.use_res_connect:
            return x + out
        else:
            return out


class MobileNet_V2(nn.Module):
    def __init__(
        self,
        num_classes: int = 1000,
        width_mult: float = 1.0,
        inverted_residual_setting: Optional[List[List[int]]] = None,
        round_nearest: int = 8,
        block: Optional[Callable[..., nn.Module]] = None,
        norm_layer: Type[BNLayer] = nn.BatchNorm2d,
        conv_layer: Type[Conv2dLayer] = nn.Conv2d,
        dropout: float = 0.2,
        input_shape: Tuple[int, int, int] = (3, 224, 224),
        is_first_conv_dense: bool = True,
        is_last_fc_dense: bool = True,
    ) -> None:
        """
        MobileNet V2 main class

        Args:
            num_classes (int): Number of classes
            width_mult (float): Width multiplier - adjusts number of channels in each layer by this amount
            inverted_residual_setting: Network structure
            round_nearest (int): Round the number of channels in each layer to be a multiple of this number
            Set to 1 to turn off rounding
            block: Module specifying inverted residual building block for mobilenet
            norm_layer: Module specifying the normalization layer to use
            dropout (float): The dropout probability

        """
        super().__init__()

        if block is None:
            block = InvertedResidual

        input_channel = 32
        last_channel = 1280

        if inverted_residual_setting is None:
            inverted_residual_setting = [
                # t, c, n, s
                [1, 16, 1, 1],
                [6, 24, 2, 2],
                [6, 32, 3, 2],
                [6, 64, 4, 2],
                [6, 96, 3, 1],
                [6, 160, 3, 2],
                [6, 320, 1, 1],
            ]

        # only check the first element, assuming user knows t,c,n,s are required
        if len(inverted_residual_setting) == 0 or len(inverted_residual_setting[0]) != 4:
            raise ValueError(
                f"inverted_residual_setting should be non-empty or a 4-element list, got {inverted_residual_setting}"
            )

        # building first layer
        input_channel = _make_divisible(input_channel * width_mult, round_nearest)
        self.last_channel = _make_divisible(last_channel * max(1.0, width_mult), round_nearest)

        features: List[nn.Module] = [
            Conv2dNormActivation(
                input_shape[0],
                input_channel,
                stride=2,
                norm_layer=nn.BatchNorm2d if is_first_conv_dense else norm_layer,
                activation_layer=nn.ReLU6,
                conv_layer=nn.Conv2d if is_first_conv_dense else conv_layer,
            )
        ]
        # building inverted residual blocks
        for t, c, n, s in inverted_residual_setting:
            output_channel = _make_divisible(c * width_mult, round_nearest)
            for i in range(n):
                stride = s if i == 0 else 1
                features.append(
                    block(
                        input_channel,
                        output_channel,
                        stride,
                        expand_ratio=t,
                        norm_layer=norm_layer,
                        conv_layer=conv_layer,
                    )
                )
                input_channel = output_channel

        # building last several layers
        features.append(
            Conv2dNormActivation(
                input_channel,
                self.last_channel,
                kernel_size=1,
                norm_layer=norm_layer,
                activation_layer=nn.ReLU6,
                conv_layer=conv_layer,
            )
        )
        # make it nn.Sequential
        self.features = nn.Sequential(*features)

        # building classifier
        if is_last_fc_dense or conv_layer != sparse.MaskedConv2d:
            # If the model is dense, we enforce the last layer to be dense
            last_fc_layer = nn.Linear
        else:
            last_fc_layer = sparse.MaskedLinear

        self.classifier = nn.Sequential(
            nn.Dropout(p=dropout), last_fc_layer(in_features=self.last_channel, out_features=num_classes)
        )

        # weight initialization
        for m in self.modules():
            if isinstance(m, (nn.Conv2d, sparse.MaskedConv2d)):
                nn.init.kaiming_normal_(m.weight, mode="fan_out")
                if m.bias is not None:
                    nn.init.zeros_(m.bias)
            elif isinstance(m, (nn.BatchNorm2d, sparse.MaskedBatchNorm2d)):
                nn.init.ones_(m.weight)
                nn.init.zeros_(m.bias)
            elif isinstance(m, nn.Linear):
                nn.init.normal_(m.weight, 0, 0.01)
                nn.init.zeros_(m.bias)

    def _forward_impl(self, x: Tensor) -> Tensor:
        # This exists since TorchScript doesn't support inheritance, so the superclass method
        # (this one) needs to have a name other than `forward` that can be accessed in a subclass
        x = self.features(x)
        # Cannot use "squeeze" as batch-size can be 1
        x = nn.functional.adaptive_avg_pool2d(x, (1, 1))
        x = torch.flatten(x, 1)

        if isinstance(self.classifier[1], sparse.MaskedLinear):
            x, _ = self.classifier(x)
        else:
            x = self.classifier(x)

        return x

    def forward(self, x: Tensor) -> Tensor:
        return self._forward_impl(x)
